smart-scheduler-api/requirements.txt
fastapi
uvicorn[standard]
beanie
numpy
openai
passlib[bcrypt]
python-jose
email-validator
dnspython
python-multipart